New: Split installation scripts into their own classes
This is the first of a series of refactors to reduce the size of
ansible_playbook.rb and use polymorphism rather than if/else
statements to accomplish the same goals.

Each operating system now has it's own class. For now, I've only moved
the installation methods to them, but the rest should follow shortly.

This has been tested by setting ansible_platform in .kitchen.yml with
the following values: debian, redhat, amazon, suse