Fix RHEL and CentOS dependencies when using pip
When installing Ansible using `pip` it will pull in cryptography as a
dependency. This needs to compile some things from source. Looks like
`libssl-devel` was added in error to the RHEL family. This should be
`openssl-devel`. Also, the `gcc` compiler needs to be available.
Depending on the image used when testing, this may not be baked in.