Fix Ansible install for Fedora 30
With the impending end-of-support for Python 2.x in Jan
2020, Fedora has started switching all Python support to
3.x. Fedora 30 has been the removal of Python 2.x packages
(see https://fedoraproject.org/wiki/Changes/Mass_Python_2_Package_Removal).
Because of this, there is no longer a python2-dnf package, so
the dnf install fails.

This command adds a check for Fedora release. If less than or
equal to 28, then use the previous install command. Starting
with 29, Fedora's Ansible is python3 based, so flip libselinux
and dnf python packages to be their python3 counterparts.